pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperliquid-ai-sim"
version = "0.1.0"
description = "AI Trading Agent for Hyperliquid with Simulation Environment"
readme = "README.md"
authors = [{name = "AI Trading Agent Developer"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "flake8",
    "mypy"
]

[project.scripts]
hypeai-sim = "src.main:main"